+++ R/ILSAdownload.R
@@ -30,7 +30,8 @@
 #' # For example, to download 'RLII' 1991 data:
 #'
 #' # Path were files will be saved
-#' output <- tempdir()
+#' dir.create(file.path(tempdir(),"ILSAdownload"))
+#' output <- file.path(tempdir(),"ILSAdownload")
 #'
 #' 
 #' # Downloading 'RLII' 1991 and unzipping files
@@ -39,7 +40,7 @@
 #'
 #' @export
 
-ILSAdownload <- function(study, year, outputdir, 
+ILSAdownload <- function(study, year, outputdir = getwd(), 
                           unzip = FALSE, maxtime = 999, 
                           quiet = FALSE, agreeLicense = FALSE){
   
@@ -84,9 +85,9 @@ ILSAdownload <- function(study, year, outputdir,
                 "\nIt should be a numeric value."))
   
   ## outputdir
-  if(!(is.vector(outputdir)&&is.character(outputdir)))
+  if(!(is.vector(outputdir)&&is.character(outputdir)&&length(outputdir)==1))
     stop(c("\nInvalid input for 'outputdir'.",
-           "\nIt should be a character vector."),call. = FALSE)
+           "\nIt should be a string."),call. = FALSE)
   
   if(!file.exists(outputdir))
     stop(c("\nInvalid input for 'outputdir'.",