+++ R/ILSAfile.info.R
@@ -7,38 +7,35 @@
 #' @returns A data frame with the number of files and MBs per population.
 #'
 #' @examples
-#' # For example, after downloading 'RLII' 1991 G4 data:
-#'
-#' # Downloading 'RLII' 1991 and unzipping files
-#' ILSAdownload(study = "RLII", year = 1991, outputdir = tempdir(), unzip = TRUE, agreeLicense = TRUE)
-#'
 #' # Path were raw 'SPSS' files are
-#' input <- file.path(tempdir(),"RLII1991_IDB_SPSS/Data")
+#' input <- system.file("extdata/reds", package = "ILSAmerge")
 #'
 #' # Get file information
 #' ILSAfile.info(inputdir = input)
 #'
 #' @export
 
-ILSAfile.info <- function(inputdir){
+ILSAfile.info <- function(inputdir = getwd()){
 
   # Checks ----
 
-  if(!(is.vector(inputdir)&&is.character(inputdir)))
+  ## inputdir
+  if(!(is.vector(inputdir)&&is.character(inputdir)&&length(inputdir)==1))
     stop(c("\nInvalid input for 'inputdir'.",
-           "\nIt should be a character vector."),call. = FALSE)
-
+           "\nIt should be a string."),call. = FALSE)
+  
   if(!file.exists(inputdir))
     stop(c("\nInvalid input for 'inputdir'.",
            "\nPath does not exist."),call. = FALSE)
 
+
   # Process & Output ----
 
-  ark <- list.files(path = inputdir,pattern = ".sav|.zsav|.SAV|.ZAV")
-  erk <- list.files(path = inputdir,pattern = ".sav|.zsav|.SAV|.ZAV",full.names = TRUE)
+  ark <- list.files(path = inputdir,pattern = ".sav|.zsav|.SAV|.ZAV",recursive = FALSE)
+  erk <- list.files(path = inputdir,pattern = ".sav|.zsav|.SAV|.ZAV",full.names = TRUE,recursive = FALSE)
 
   if(length(ark)==0)
-    stop(c("\nNo files found."),call. = FALSE)
+    stop(c("\nNo ILSA raw files found."),call. = FALSE)
 
 
 
@@ -54,13 +51,13 @@ ILSAfile.info <- function(inputdir){
   stu <- unlist(lapply(ext,function(i) i[[1]]))
   ext <- unlist(lapply(ext,function(i) i[[2]]))
 
-  popstu <- paste0(pop,stu)
+  popstu <- toupper(paste0(pop,stu))
   upopstu <- unique(popstu)
 
   siz <- file.size(erk)
   siz <- stats::aggregate(siz,by = list(popstu),sum)
 
-  out <- as.data.frame(table(popstu))
+  out <- as.data.frame(table(popstu),stringsAsFactors = FALSE)
   colnames(out) <- c('Population','Files')
   out <- cbind.data.frame(out,MB = round(siz[,2]/1000/1000,1))