+++ R/ILSAmerge.R
@@ -3,6 +3,10 @@
 #' Merges 'SPSS' data from different International Large-Scale Assessments (ILSA).
 #' This function has been tested to behave correctly for: 'TIMSS', 'TIMSS Advanced', 
 #' 'PIRLS', 'ICCS', 'ICILS', 'CIVED', 'REDS', 'RLII', and 'SITES' (2006).
+#' 
+#' For files merged within R it will also add country information where needed. 
+#' Country information will be retrieved from 'GitHub' if possible. If not, it will
+#' use the package internal data. 
 #'
 #' @param inputdir a string indicating the path were ILSA 'SPSS' files are stored.
 #' @param outputdir the directory where the merged data will be saved.
@@ -18,6 +22,8 @@
 #' If set to \code{NULL}, no limit will be used and all files will be merged
 #' within R. If speed is a problem, we recommend
 #' that this number should not be over \code{200} and merge the rest in 'SPSS'.
+#' Beware that some ILSA will have files with different columns and this could
+#' cause some 'SPSS' syntaxes to fail. If this happens, merge through \code{R}.
 #' @param MBlistlimit a numerical value indicating the allowed limit of the
 #' combined storage of the files of one type for merging through a list.
 #' Values over the limit will be merged through a matrix, which will be slower
@@ -30,18 +36,14 @@
 #' @returns Saves merged ILSA data or \code{.sps} syntax for merging ILSA data.
 #'
 #' @examples
-#' # For example, after downloading 'RLII' 1991 G4 data:
-#'
-#' # Downloading 'RLII' 1991 and unzipping files
-#' ILSAdownload(study = "RLII", year = 1991, outputdir = tempdir(), unzip = TRUE, agreeLicense = TRUE)
-#' 
 #' # Path were raw 'SPSS' files are
-#' input <- file.path(tempdir(),"RLII1991_IDB_SPSS/Data")
+#' input <- system.file("extdata/reds", package = "ILSAmerge")
 #' 
 #' # Path were merged files will be saved
-#' output <- file.path(tempdir(),"RLII1991_IDB_SPSS")
+#' dir.create(file.path(tempdir(),"ILSAmerge"))
+#' output <- file.path(tempdir(),"ILSAmerge")
 #' 
-#' # Merging 'RLII' 1991, as .rds file
+#' # Merging 'REDS' 2021, as .rds file
 #' ILSAmerge(inputdir = input, outputdir = output, filetype = "rds", quiet = FALSE)
 #'
 #' @export
@@ -49,27 +51,27 @@
 
 
 
-
-
-ILSAmerge <- function(inputdir, outputdir, population = NULL,
+ILSAmerge <- function(inputdir = getwd(), outputdir = getwd(), population = NULL,
                       filetype = c("rds", "zsav", "sav"),
                       MBlimit = NULL, MBlistlimit = 200, SPSSlimit = 50,
                       quiet = FALSE){
 
   # Checks ----
 
-  if(!(is.vector(inputdir)&&is.character(inputdir)))
+  ## inputdir
+  if(!(is.vector(inputdir)&&is.character(inputdir)&&length(inputdir)==1))
     stop(c("\nInvalid input for 'inputdir'.",
-           "\nIt should be a character vector."),call. = FALSE)
-
+           "\nIt should be a string."),call. = FALSE)
+  
   if(!file.exists(inputdir))
     stop(c("\nInvalid input for 'inputdir'.",
            "\nPath does not exist."),call. = FALSE)
-
-  if(!(is.vector(outputdir)&&is.character(outputdir)))
+  
+  ## outputdir
+  if(!(is.vector(outputdir)&&is.character(outputdir)&&length(outputdir)==1))
     stop(c("\nInvalid input for 'outputdir'.",
-           "\nIt should be a character vector."),call. = FALSE)
-
+           "\nIt should be a string."),call. = FALSE)
+  
   if(!file.exists(outputdir))
     stop(c("\nInvalid input for 'outputdir'.",
            "\nPath does not exist."),call. = FALSE)
@@ -109,8 +111,8 @@ ILSAmerge <- function(inputdir, outputdir, population = NULL,
 
   filetype <- match.arg(filetype,c("rds", "zsav", "sav"))
 
-  ark <- list.files(path = inputdir,pattern = ".sav|.zsav|.SAV|.ZSAV")
-  erk <- list.files(path = inputdir,pattern = ".sav|.zsav|.SAV|.ZSAV",full.names = TRUE)
+  ark <- list.files(path = inputdir,pattern = ".sav|.zsav|.SAV|.ZSAV",recursive = FALSE)
+  erk <- list.files(path = inputdir,pattern = ".sav|.zsav|.SAV|.ZSAV",full.names = TRUE,recursive = FALSE)
 
   pop <- substr(ark,1,3)
   cou <- substr(ark,4,6)
@@ -128,7 +130,7 @@ ILSAmerge <- function(inputdir, outputdir, population = NULL,
 
 
 
-  popstu <- paste0(pop,stu)
+  popstu <- toupper(paste0(pop,stu))
   upopstu <- sort(unique(popstu))
 
   if(!is.null(population)){
@@ -148,7 +150,7 @@ ILSAmerge <- function(inputdir, outputdir, population = NULL,
 
 
   if(!quiet){
-    cat(paste0(length(ark)," files detected. Merging into ",
+    cat(paste0(sum(popstu%in%upopstu)," files detected. Merging into ",
                length(upopstu)," files.\n"))
   }
   
@@ -166,6 +168,31 @@ ILSAmerge <- function(inputdir, outputdir, population = NULL,
 
 
   ### Merge ----
+  
+  # Add country
+  
+  where <- "https://raw.githubusercontent.com/dopatendo/ILSAmerge/refs/heads/main/data/ILSAcou.csv"
+  
+  
+  where <- suppressWarnings(try(utils::read.csv(where),silent = TRUE))
+  
+  if("try-error"%in%class(where)){
+    warning(paste0("Could not read country information from 'GitHub'.",
+                "\nInternal data will be used for adding country labels.",
+                "\nPlease be aware, these data may not be the lastest one."),call. = FALSE)
+    
+    ILSAcou <- utils::read.csv(file.path(system.file("extdata/ilsainfo", package = "ILSAmerge"),"ILSAcou.csv"))
+  }else{
+    ILSAcou <- where
+  }
+  
+  
+  ILSAcou <- ILSAcou[ILSAcou$N3code!=0,]
+  couL <- ILSAcou$IEAcode
+  names(couL) <- ILSAcou$Name
+  couLS <- sort(couL[ILSAcou$toLabel%in%1])
+  couL <- sort(couL)
+  
 
 
   ptime <- proc.time()
@@ -177,6 +204,7 @@ ILSAmerge <- function(inputdir, outputdir, population = NULL,
       cat(paste0("Merging ",upopstu[i],". Type ",i," of ",length(upopstu),".\n"))
     }
 
+    coui <- toupper(cou[popstu%in%upopstu[i]])
     erki <- erk[popstu%in%upopstu[i]]
     mbs <- inf[,3][inf[,1]%in%upopstu[i]]
 
@@ -184,14 +212,14 @@ ILSAmerge <- function(inputdir, outputdir, population = NULL,
 
       if(mbs<=(MBlistlimit+0.01)){
 
-        out <- try(.mergebylist(files = erki,verbose = !quiet),silent = TRUE)
+        out <- try(.mergebylist(files = erki,verbose = !quiet, cous = coui, couL = couL, couLS = couLS),silent = TRUE)
 
         if("try-error"%in%class(out)){
-          out <- .mergebymatrix(files = erki,verbose = !quiet)
+          out <- .mergebymatrix(files = erki,verbose = !quiet, couL = couL, couLS = couLS)
         }
 
       }else{
-        out <- .mergebymatrix(files = erki,verbose = !quiet)
+        out <- .mergebymatrix(files = erki,verbose = !quiet, couL = couL, couLS = couLS)
       }
 
       if(filetype%in%"zsav"){
@@ -244,7 +272,9 @@ ILSAmerge <- function(inputdir, outputdir, population = NULL,
 
 
 
-.mergebymatrix <- function(files,verbose = TRUE){
+.mergebymatrix <- function(files,verbose = TRUE, couL, couLS){
+  
+
   # first file to load attributes
 
   if(verbose)
@@ -277,13 +307,20 @@ ILSAmerge <- function(inputdir, outputdir, population = NULL,
 
 
   out1 <- try(haven::read_spss(file = files[whtoload], user_na = TRUE, col_select = NULL,
-                               skip = 0, n_max = 1, .name_repair = "unique"),
+                               skip = 0, n_max = 0, .name_repair = "unique"),
               silent = TRUE)
   if("try-error"%in%class(out1)){
     out1 <- haven::read_sav(file = files[whtoload], user_na = TRUE, col_select = NULL,
                             skip = 0, n_max = 0, .name_repair = "unique",
                             encoding = 'latin1')
+    
+
   }
+  colnames(out1) <- toupper(colnames(out1))
+  
+  atrs <- lapply(1:ncol(out1),function(h){
+    attributes(out1[,h,drop = TRUE])
+  })
 
 
 
@@ -300,7 +337,12 @@ ILSAmerge <- function(inputdir, outputdir, population = NULL,
 
     nrow(load)
   })
-
+  
+  COU <- substr(basename(files[unlist(last)>0]),4,6)
+  COUWD <- last[unlist(last)>0]
+  COU <- unlist(lapply(1:length(COU),function(co) rep(COU[co],COUWD[[co]])))
+  hasdata <- sapply(last,function(j) j!=0)
+  
 
   # last <- lapply(files[-1],function(j){
   #   nrow(haven::read_spss(file = j,col_select = 'IDCNTRY'))
@@ -313,7 +355,7 @@ ILSAmerge <- function(inputdir, outputdir, population = NULL,
   out <- (matrix(NA,nrow = max(last),ncol = ncol(out1)))
   colnames(out) <- toupper(colnames(out1))
 
-
+  
 
   for(j in 1:length(files)){
 
@@ -323,42 +365,48 @@ ILSAmerge <- function(inputdir, outputdir, population = NULL,
     if(verbose)
       cat(paste0("Merging dataset ",j," of ",length(files),".\n"))
 
-
-
-    unt <- try(haven::read_spss(file = files[j],
-                                user_na = TRUE,
-                                col_select = NULL,
-                                skip = 0,
-                                n_max = Inf,
-                                .name_repair = "unique"),
-               silent = TRUE)
-
-    if("try-error"%in%class(unt)){
-      unt <- haven::read_sav(file = files[j],
-                              user_na = TRUE,
-                              col_select = NULL,
-                              skip = 0,
-                              n_max = Inf,
-                              .name_repair = "unique",
-                              encoding = "latin1")
+    if(hasdata[j]){
+      unt <- try(haven::read_spss(file = files[j],
+                                  user_na = TRUE,
+                                  col_select = NULL,
+                                  skip = 0,
+                                  n_max = Inf,
+                                  .name_repair = "unique"),
+                 silent = TRUE)
+      
+      if("try-error"%in%class(unt)){
+        unt <- haven::read_sav(file = files[j],
+                               user_na = TRUE,
+                               col_select = NULL,
+                               skip = 0,
+                               n_max = Inf,
+                               .name_repair = "unique",
+                               encoding = "latin1")
+      }
+      
+      
+      cunt <- toupper(colnames(unt))
+      unt <- lapply(1:ncol(unt), function(X) {
+        as.vector(unt[, X, drop = TRUE])
+      })
+      unt <- do.call(cbind.data.frame, unt)
+      
+      mts <- match(cunt,colN)
+      
+      for(k in 1:length(cunt)){
+        out[first[j]:last[j],mts[k]] <- unt[,k]
+      }
+      
+      # out[first[j]:last[j],match(cunt,colN)] <- unt
+      
     }
 
 
 
 
-    cunt <- toupper(colnames(unt))
-    unt <- lapply(1:ncol(unt), function(X) {
-      as.vector(unt[, X, drop = TRUE])
-    })
-    unt <- do.call(cbind.data.frame, unt)
 
-    mts <- match(cunt,colN)
-    
-    for(k in 1:length(cunt)){
-      out[first[j]:last[j],mts[k]] <- unt[,k]
-    }
-    
-    # out[first[j]:last[j],match(cunt,colN)] <- unt
+
+ 
     
   }
     # })
@@ -366,17 +414,69 @@ ILSAmerge <- function(inputdir, outputdir, population = NULL,
 
 
 
-  colnames(out) <- colnames(out1)
+  # colnames(out) <- colnames(out1)
 
+  # Fix date
 
-  out <- rbind(out1,out)[-1,]
+  
+  isdate <- which(sapply(1:ncol(out1),function(k) "Date"%in%class(out1[,k,drop = TRUE])))
+  
+  if(length(isdate)>0){
+    for(d in isdate){
+      out[,d]<- as.Date.numeric(out[,d],origin="1970-01-01")
+    }
+    
+  }
+  
+  out <- as.data.frame(out)
+  
+  # REPAIR attributes
+  
+    class(out) <- class(out1)
+    for(h in 1:ncol(out)){
+      attributes(out[,h,drop = TRUE]) <- atrs[[h]]
+    }
+    
+
+  
+  
+  
+  # out <- rbind(out1,out)[-1,]
+  
+
+  
+  # add labels to IDCNTRY
+  
+  # if(is.character(out$IDCNTRY)){
+  #   out$IDCNTRY <- as.numeric(out$IDCNTRY)
+  # }
+  attr(out$IDCNTRY,'labels') <- couLS
+
+  
+  # Add country string
+  if(!"IDCNTRY_STR"%in%colnames(out)){
+    cl <- class(out)
+    idstr <- as.numeric(as.character(out$IDCNTRY))
+    out <- cbind(IDCNTRY_STR = names(couL)[match(idstr,couL)], out)
+    class(out) <- cl
+  }
+  
+  if(!"CNTRY"%in%colnames(out)){
+    cl <- class(out)
+    # mtc <- match(as.numeric(out$IDCNTRY),ILSAcou$IEAcode)
+    # CNTRY <- rep(NA,length(mtc))
+    # CNTRY[!is.na(mtc)] <- ILSAcou$N3code[!is.na(mtc)]
+    out <- cbind(CNTRY = COU, out)
+    class(out) <- cl
+  }
+
 
   return(out)
 
 }
 
 
-.mergebylist <- function(files,verbose = TRUE){
+.mergebylist <- function(files,verbose = TRUE, cous, couL, couLS){
 
 
 
@@ -388,6 +488,27 @@ ILSAmerge <- function(inputdir, outputdir, population = NULL,
     outj <- haven::read_spss(file = files[j], user_na = TRUE, col_select = NULL,
                      skip = 0, n_max = Inf, .name_repair = "unique")
     colnames(outj) <- toupper(colnames(outj))
+    
+    
+    
+    
+    # add labels to IDCNTRY
+    attr(outj$IDCNTRY,'labels') <- couLS
+    
+    # Add country string
+    if(!"IDCNTRY_STR"%in%colnames(outj)){
+      cl <- class(outj)
+      outj <- cbind(IDCNTRY_STR = names(couL)[match(as.numeric(outj$IDCNTRY),couL)], outj)
+      class(outj) <- cl
+    }
+    
+    if(!"CNTRY"%in%colnames(outj)){
+      cl <- class(outj)
+      outj <- cbind(CNTRY = cous[j], outj)
+      class(outj) <- cl
+    }
+    
+    
     outj
   })
 
@@ -396,6 +517,5 @@ ILSAmerge <- function(inputdir, outputdir, population = NULL,
   return(out)
 
 }
-