+++ inst/doc/IdentifyILSAfiles.R
@@ -1,0 +1,13 @@
+## ----include = FALSE----------------------------------------------------------
+knitr::opts_chunk$set(
+  collapse = TRUE
+)
+
+## ----setup, echo = FALSE------------------------------------------------------
+library(ILSAmerge)
+
+## -----------------------------------------------------------------------------
+
+ILSAfile.info(inputdir = system.file("extdata/reds", package = "ILSAmerge"))
+
+