+++ NEWS.md
@@ -1,13 +1,32 @@
-# ILSAmerge 1.0.0
- - CRAN submission.
+# ILSAmerge 1.2.5
+- Added `combineStudents()` for combining achievement and  background questionnaires.
+- Added `addSchools()` for adding school data.
+- Added labels to countries and creates variables with country codes and country names.
+
+### Remaining issues
+- No unzip of PISA and TALIS downloads.
+
+
+
+# ILSAmerge 1.0.5
+- Added `ILSArename()` for creating comprehensible file names.
+- Fixed a problem caused by PIRLS 2021 having an empty dataset.
+- Fixed a problem with merging dates for TIMSS 2011. 
+- Fixed a problem originated from PIRLS 2001 with its multiple capitalization styles. 
+
+
+# ILSAmerge 1.0.1
+- Added extensive new documentation.
+- Added `availableILSA()` to check which ILSA can be downloaded.
 
 # ILSAmerge 1.0.0
-- Added PISA and TALIS to `ILSAdownload`.
-- `ILSAdownload` connects to GitHub to check for most recent links.
+- CRAN submission.
+- Added PISA and TALIS to `ILSAdownload()`.
+- `ILSAdownload()` connects to GitHub to check for most recent links.
 
 # ILSAmerge 0.8.4
 - Fixed internal argument checks.
-- Added `ILSAdownload` for downloading ILSA SPSS datasets.
+- Added `ILSAdownload()` for downloading ILSA 'SPSS' datasets.
 
 # ILSAmerge 0.8.1
  - Added `justload()` to load files in a list but not merge them or save them.